% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_csv}
\alias{append_csv}
\title{Append a csv file}
\usage{
append_csv(file, print = FALSE)
}
\arguments{
\item{file}{the file to append the result to.}

\item{print}{Should the result be printed to the console?}
}
\description{
Append a csv file
}
