% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_b}
\alias{to_b}
\alias{to_kib}
\alias{to_mib}
\alias{to_gib}
\title{Convert Byte notation to Numeric}
\usage{
to_b(units)

to_kib(units)

to_mib(units)

to_gib(units)
}
\arguments{
\item{units}{A character vector to convert to bytes}
}
\description{
These function turnsbyte notation to numeric using a
byte conversion. Useful for turning the \code{memUsage} column
to numeric.
}
\note{
The following table are used, based on Google Unit converter:
\subsection{bytes}{
\itemize{
\item "b" = 1,
\item "kib" = 125,
\item "mib" = 1049000,
\item "gib" = 1.074e+9
\item 
}
}

\subsection{kib}{
\itemize{
\item "b" = 1/1024,
\item "kib" = 1,
\item "mib" = 1024,
\item "gib" = 1049000
}
}

\subsection{mib}{
\itemize{
\item "b" = 1/1.049e+6,
\item "kib" = 1/1024,
\item "mib" = 1,
\item "gib" = 1024
}
}

\subsection{mib}{
\itemize{
\item "b" = 1/1.074e+9,
\item "kib" = 1/1.049e+6,
\item "mib" = 1/1024,
\item "gib" = 1
}
}
}
\examples{
un <- c("55.12kiB","12.09MiB","21.75GiB","467.2MiB","42.91MiB")
to_b(un)
to_kib(un)
to_mib(un)
to_gib(un)
}
