% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{append_csv}
\alias{append_csv}
\alias{read_appended_csv}
\title{Append a csv file and read it in R}
\usage{
append_csv(file, print = FALSE)

read_appended_csv(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an EOF signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{print}{Should the result be printed to the console?}
}
\description{
\code{append_csv()} returns a function that can the be used to
append a csv with a dataframe content.
The returned function is write.table() with the following
parameters : \verb{append = TRUE, col.names = FALSE, row.names = FALSE, sep = ","}.
}
\details{
\code{read_appended_csv()} is a wrapper around \code{read.csv()},.
so please use that function for more control over the
reading.
}
\examples{
if (interactive()){
    if (dockerstats_available()) {
        fls <- tempfile(fileext = ".csv")
        write_it <- append_csv(fls)
        write_it(dockerstats())
        read_appended_csv(fls)
    }
}
}
