% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\docType{data}
\name{docker_stats_names}
\alias{docker_stats_names}
\title{Names of the outputs}
\format{
An object of class \code{character} of length 14.
}
\usage{
docker_stats_names
}
\description{
Names of the outputs
}
\keyword{datasets}
