% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dockerstats}
\alias{dockerstats}
\title{Run docker stats}
\usage{
dockerstats(..., all = FALSE, extra = "")
}
\arguments{
\item{...}{Subset of containers to get stats from. If empty, all running containers are looked for.}

\item{all}{Show all containers (default shows just running).}

\item{extra}{Extra information to add in the "extra" column.}
}
\value{
A data.frame
}
\description{
Run docker stats and returns a data.frame
}
