% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{dockerstats_available}
\alias{dockerstats_available}
\title{Is \verb{docker stats} available?}
\usage{
dockerstats_available()
}
\value{
\code{invisible(TRUE)} if \code{system("docker stats")} return 0, \code{invisible(FALSE)} otherwise.
}
\description{
This function tries to lauch \verb{docker stats} on the machine.
If the status code return by this command is not \code{0} (success),
this function will print a message and return FALSE.
If the return code is 0, this function invisibly return \code{TRUE}.
}
\examples{
if (interactive()){
    dockerstats_available()
}
}
