% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_recurse.R
\name{dockerstats_recurse}
\alias{dockerstats_recurse}
\title{Call dockerstats() recursively}
\usage{
dockerstats_recurse(..., all = FALSE, every = 1, callback = print)
}
\arguments{
\item{...}{Subset of containers to get stats from. If empty, all running containers are looked for.}

\item{all}{Show all containers (default shows just running).}

\item{every}{Loop duration}

\item{callback}{function to call on each loop, with the first argument being the result of dockerstats().}
}
\description{
Call dockerstats() recursively
}
