% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_b}
\alias{to_b}
\alias{to_kib}
\alias{to_mib}
\alias{to_gib}
\title{Convert Byte notation to Numeric}
\usage{
to_b(units)

to_kib(units)

to_mib(units)

to_gib(units)
}
\arguments{
\item{units}{A character vector to convert to bytes}
}
\value{
A character vector of size converted from the input. The
\code{attr("unit")} contains the unit of the result.
}
\description{
These function turnsbyte notation to numeric using a
byte conversion. Useful for turning the \code{memUsage} column
to numeric.
}
\note{
The following table are used, based on Google Unit converter:
\subsection{bytes}{
\itemize{
\item "1b" = 1 bytes,
\item "1kib" = 1024 bytes,
\item "1mib" = 1049000 bytes,
\item "1gib" = 1.074e+9 bytes
\item 
}
}

\subsection{kib}{
\itemize{
\item "1b" = 1/1024 kib,
\item "1kib" = 1 kib,
\item "1mib" = 1024 kib,
\item "1gib" = 1049000 kib
}
}

\subsection{mib}{
\itemize{
\item "1 b" = 1/1.049e+6 mib,
\item "1 kib" = 1/1024 mib,
\item "1 mib" = 1 mib,
\item "1 gib" = 1024 mib
}
}

\subsection{gib}{
\itemize{
\item "1b" = 1/1.074e+9 gib,
\item "1kib" = 1/1.049e+6 gib,
\item "1mib" = 1/1024 gib,
\item "1gib" = 1 gib
}
}
}
\examples{
un <- c("55.12kiB","12.09MiB","21.75GiB","467.2MiB","42.91MiB")
to_b(un)
to_kib(un)
to_mib(un)
to_gib(un)
}
