% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerstats-package.R
\docType{package}
\name{dockerstats-package}
\alias{dockerstats}
\alias{dockerstats-package}
\title{dockerstats: A Wrapper Around 'docker stats'}
\description{
A wrapper around docker stats..
}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

}
\keyword{internal}
