% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerstats-package.R
\docType{package}
\name{dockerstats-package}
\alias{dockerstats-package}
\title{dockerstats package}
\description{
A wrapper around the \code{dockerstats} command to analyse doker data from R.
}
\author{
colin \href{mailto:contact@colinfay.me}{contact@colinfay.me}
}
